+++ relay-server/src/testutils.rs
@@ -13,7 +13,6 @@ use relay_system::Addr;
 use relay_test::mock_service;
 
 use crate::envelope::{Envelope, Item, ItemType};
-use crate::extractors::RequestMeta;
 use crate::metrics::{MetricOutcomes, MetricStats};
 #[cfg(feature = "processing")]
 use crate::service::create_redis_pools;
@@ -101,18 +100,6 @@ pub fn new_envelope<T: Into<String>>(with_dsc: bool, transaction_name: T) -> Box
     let item3 = Item::new(ItemType::Attachment);
     envelope.add_item(item3);
 
-    envelope
-}
-
-pub fn empty_envelope() -> Box<Envelope> {
-    empty_envelope_with_dsn("e12d836b15bb49d7bbf99e64295d995b")
-}
-
-pub fn empty_envelope_with_dsn(dsn: &str) -> Box<Envelope> {
-    let dsn = format!("https://{dsn}:@sentry.io/42").parse().unwrap();
-
-    let mut envelope = Envelope::from_request(Some(EventId::new()), RequestMeta::new(dsn));
-    envelope.add_item(Item::new(ItemType::Event));
     envelope
 }